Revert "updated secrets installer to take in named parameters for optional handling of variables"
This reverts commit 958829b5b1d94c8b8c59be21c1ea2c9bf1a343aa.

+++ installer/update_secrets.sh
@@ -1,32 +1,15 @@
 #!/bin/bash -e
-while getopts ':e:p:t:' OPTION; do
-    case $OPTION in
-        e)
-          # Environment name
-          ENVIRONMENT="$OPTARG"
-          echo "The name of the environment is $ENVIRONMENT" ;;
-        p)
-          # Use 1Password to bootstrap secrets or not.  Set to 1 to enable
-          USE_ONE_PASSWORD="$OPTARG"
-          echo " one password is $USE_ONE_PASSWORD" ;;
-        t)
-          #1 Password Connect token
-          OP_CONNECT_TOKEN="$OPTARG"
-          echo " one password is $USE_ONE_PASSWORD" ;;
-        *)
-          # Printing error message
-          echo echo "Usage: $(basename $0) [-e] [-p] [-t]"
-          exit 1
-          ;;
-    esac
-done
-
+ENVIRONMENT=$1
 
 export OP_CONNECT_HOST=https://roundtable.lsst.codes/1password
 export VAULT_DOC_UUID=`yq -r .onepasswordUuid ../environments/values.yaml`
-set ${VAULT_ADDR:=https://vault.lsst.codes}
-set ${VAULT_TOKEN=`./vault_key.py $ENVIRONMENT write`}
+export VAULT_ADDR=https://vault.lsst.codes
+export VAULT_TOKEN=`./vault_key.py $ENVIRONMENT write`
 
+if [ -z "$OP_CONNECT_TOKEN" ]; then
+    echo 'OP_CONNECT_TOKEN must be set to a 1Password Connect token' >&2
+    exit 1
+fi
 
 echo "Clear out any existing secrets"
 rm -rf secrets
@@ -34,13 +17,8 @@ rm -rf secrets
 echo "Reading current secrets from vault"
 ./read_secrets.sh $ENVIRONMENT
 
-if [ "${USE_ONE_PASSWORD}" == 1 ]; then
-    echo "Generating missing secrets with values from 1Password"
-    ./generate_secrets.py $ENVIRONMENT --op
-else
-    echo "Open Password disabled.  Generating missing secrets"
-    ./generate_secrets.py $ENVIRONMENT 
-fi
+echo "Generating missing secrets with values from 1Password"
+./generate_secrets.py $ENVIRONMENT --op
 
 echo "Writing secrets to vault"
 ./write_secrets.sh $ENVIRONMENT
